Main Files Upload
These are the main files based on the initial UML analysis of how the robot will function. Components are grouped into 3 objects based on object oriented programming: Robot Arm, MiniMaxAI and TicTacToe Board represent the 3 classes of this project. Current code has a working AI, a Tic Tac Toe board capable of sensing hall effect sensors and sending that information to the AI via a 2D array and the robot arm is only capable of manual controlling servos with potentiometers for now.